+++ NEWS.md
@@ -12,6 +12,8 @@
 - Support content search API with the `search_content()` function. (#272)
 - New `search_content()` function which lets you search and filter content items
   on the Connect server. (#272, #447)
+- New `lock_content()` and `unlock_content()` functions for locking and unlocking
+  content items. (#453)
 
 # connectapi 0.8.0